+++ docs/api/web-frame-main.md
@@ -189,6 +189,67 @@ When executed on a video media element, copies the frame at (x, y) to the clipbo
 
 When executed on a video media element, shows a save dialog and saves the frame at (x, y) to disk.
 
+#### `frame.printToPDF(options)`
+
+* `options` Object
+  * `landscape` boolean (optional) - Paper orientation.`true` for landscape, `false` for portrait. Defaults to false.
+  * `displayHeaderFooter` boolean (optional) - Whether to display header and footer. Defaults to false.
+  * `printBackground` boolean (optional) - Whether to print background graphics. Defaults to false.
+  * `scale` number(optional)  - Scale of the webpage rendering. Defaults to 1.
+  * `pageSize` string | Size (optional) - Specify page size of the generated PDF. Can be `A0`, `A1`, `A2`, `A3`,
+  `A4`, `A5`, `A6`, `Legal`, `Letter`, `Tabloid`, `Ledger`, or an Object containing `height` and `width` in inches. Defaults to `Letter`.
+  * `margins` Object (optional)
+    * `top` number (optional) - Top margin in inches. Defaults to 1cm (~0.4 inches).
+    * `bottom` number (optional) - Bottom margin in inches. Defaults to 1cm (~0.4 inches).
+    * `left` number (optional) - Left margin in inches. Defaults to 1cm (~0.4 inches).
+    * `right` number (optional) - Right margin in inches. Defaults to 1cm (~0.4 inches).
+  * `pageRanges` string (optional) - Page ranges to print, e.g., '1-5, 8, 11-13'. Defaults to the empty string, which means print all pages.
+  * `headerTemplate` string (optional) - HTML template for the print header. Should be valid HTML markup with following classes used to inject printing values into them: `date` (formatted print date), `title` (document title), `url` (document location), `pageNumber` (current page number) and `totalPages` (total pages in the document). For example, `<span class=title></span>` would generate span containing the title.
+  * `footerTemplate` string (optional) - HTML template for the print footer. Should use the same format as the `headerTemplate`.
+  * `preferCSSPageSize` boolean (optional) - Whether or not to prefer page size as defined by css. Defaults to false, in which case the content will be scaled to fit the paper size.
+  * `generateTaggedPDF` boolean (optional) _Experimental_ - Whether or not to generate a tagged (accessible) PDF. Defaults to false. As this property is experimental, the generated PDF may not adhere fully to PDF/UA and WCAG standards.
+  * `generateDocumentOutline` boolean (optional) _Experimental_ - Whether or not to generate a PDF document outline from content headers. Defaults to false.
+
+Returns `Promise<Buffer>` - Resolves with the generated PDF data.
+
+Prints the frame's web page as PDF.
+
+Unlike [`webContents.printToPDF`](web-contents.md#contentsprinttopdfoptions),
+this method prints only the contents of the frame it is called on. This can be
+used to print an individual `<iframe>` from the main process.
+
+The `landscape` will be ignored if `@page` CSS at-rule is used in the web page.
+
+An example of printing an iframe to PDF:
+
+```js
+const { app, BrowserWindow } = require('electron')
+
+const fs = require('node:fs')
+const os = require('node:os')
+const path = require('node:path')
+
+app.whenReady().then(() => {
+  const win = new BrowserWindow()
+  win.loadFile('page-with-iframe.html')
+
+  win.webContents.on('did-finish-load', () => {
+    const pdfPath = path.join(os.homedir(), 'Desktop', 'iframe.pdf')
+    const iframe = win.webContents.mainFrame.frames[0]
+    iframe.printToPDF({}).then(data => {
+      fs.writeFile(pdfPath, data, (error) => {
+        if (error) throw error
+        console.log(`Wrote PDF successfully to ${pdfPath}`)
+      })
+    }).catch(error => {
+      console.log(`Failed to write PDF to ${pdfPath}: `, error)
+    })
+  })
+})
+```
+
+See [Page.printToPdf](https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-printToPDF) for more information.
+
 ### Instance Properties
 
 #### `frame.ipc` _Readonly_
