+++ filenames.auto.gni
@@ -658,6 +658,7 @@ auto_filenames = {
     "lib/browser/ipc-main-internal.ts",
     "lib/browser/message-port-main.ts",
     "lib/browser/parse-features-string.ts",
+    "lib/browser/print-to-pdf.ts",
     "lib/browser/rpc-server.ts",
     "lib/browser/web-view-events.ts",
     "lib/common/api/module-list.ts",
