+++ lib/browser/api/web-contents.ts
@@ -6,6 +6,7 @@ import {
 import { IpcMainImpl } from '@electron/internal/browser/ipc-main-impl';
 import * as ipcMainUtils from '@electron/internal/browser/ipc-main-internal-utils';
 import { parseFeatures } from '@electron/internal/browser/parse-features-string';
+import { printToPDF } from '@electron/internal/browser/print-to-pdf';
 import * as deprecate from '@electron/internal/common/deprecate';
 import { IPC_MESSAGES } from '@electron/internal/common/ipc-messages';
 
@@ -19,11 +20,6 @@ import * as url from 'url';
 // before the webContents module.
 // eslint-disable-next-line no-unused-expressions
 session;
-
-let nextId = 0;
-const getNextId = function () {
-  return ++nextId;
-};
 
 // Stock page sizes
 const PDFPageSizes: Record<string, ElectronInternal.MediaSize> = {
@@ -88,20 +84,6 @@ const PDFPageSizes: Record<string, ElectronInternal.MediaSize> = {
     name: 'ISO_A6',
     width_microns: 105000
   }
-} as const;
-
-const paperFormats: Record<string, ElectronInternal.PageSize> = {
-  letter: { width: 8.5, height: 11 },
-  legal: { width: 8.5, height: 14 },
-  tabloid: { width: 11, height: 17 },
-  ledger: { width: 17, height: 11 },
-  a0: { width: 33.1, height: 46.8 },
-  a1: { width: 23.4, height: 33.1 },
-  a2: { width: 16.54, height: 23.4 },
-  a3: { width: 11.7, height: 16.54 },
-  a4: { width: 8.27, height: 11.7 },
-  a5: { width: 5.83, height: 8.27 },
-  a6: { width: 4.13, height: 5.83 }
 } as const;
 
 // The minimum micron size Chromium accepts is that where:
@@ -199,81 +181,8 @@ WebContents.prototype.executeJavaScriptInIsolatedWorld = async function (worldId
   );
 };
 
-function checkType<T>(value: T, type: 'number' | 'boolean' | 'string' | 'object', name: string): T {
-  // eslint-disable-next-line valid-typeof
-  if (typeof value !== type) {
-    throw new TypeError(`${name} must be a ${type}`);
-  }
-
-  return value;
-}
-
-function parsePageSize(pageSize: string | ElectronInternal.PageSize) {
-  if (typeof pageSize === 'string') {
-    const format = paperFormats[pageSize.toLowerCase()];
-    if (!format) {
-      throw new Error(`Invalid pageSize ${pageSize}`);
-    }
-
-    return { paperWidth: format.width, paperHeight: format.height };
-  } else if (typeof pageSize === 'object') {
-    if (typeof pageSize.width !== 'number' || typeof pageSize.height !== 'number') {
-      throw new TypeError('width and height properties are required for pageSize');
-    }
-
-    return { paperWidth: pageSize.width, paperHeight: pageSize.height };
-  } else {
-    throw new TypeError('pageSize must be a string or an object');
-  }
-}
-
-// Translate the options of printToPDF.
-
-const printToPDFQueues = new WeakMap<Electron.WebContents, Promise<unknown>>();
 WebContents.prototype.printToPDF = async function (options) {
-  const margins = checkType(options.margins ?? {}, 'object', 'margins');
-  const pageSize = parsePageSize(options.pageSize ?? 'letter');
-
-  const { top, bottom, left, right } = margins;
-  const validHeight = [top, bottom].every((u) => u === undefined || u <= pageSize.paperHeight);
-  const validWidth = [left, right].every((u) => u === undefined || u <= pageSize.paperWidth);
-
-  if (!validHeight || !validWidth) {
-    throw new Error('margins must be less than or equal to pageSize');
-  }
-
-  const printSettings = {
-    requestID: getNextId(),
-    landscape: checkType(options.landscape ?? false, 'boolean', 'landscape'),
-    displayHeaderFooter: checkType(options.displayHeaderFooter ?? false, 'boolean', 'displayHeaderFooter'),
-    headerTemplate: checkType(options.headerTemplate ?? '', 'string', 'headerTemplate'),
-    footerTemplate: checkType(options.footerTemplate ?? '', 'string', 'footerTemplate'),
-    printBackground: checkType(options.printBackground ?? false, 'boolean', 'printBackground'),
-    scale: checkType(options.scale ?? 1.0, 'number', 'scale'),
-    marginTop: checkType(margins.top ?? 0.4, 'number', 'margins.top'),
-    marginBottom: checkType(margins.bottom ?? 0.4, 'number', 'margins.bottom'),
-    marginLeft: checkType(margins.left ?? 0.4, 'number', 'margins.left'),
-    marginRight: checkType(margins.right ?? 0.4, 'number', 'margins.right'),
-    pageRanges: checkType(options.pageRanges ?? '', 'string', 'pageRanges'),
-    preferCSSPageSize: checkType(options.preferCSSPageSize ?? false, 'boolean', 'preferCSSPageSize'),
-    generateTaggedPDF: checkType(options.generateTaggedPDF ?? false, 'boolean', 'generateTaggedPDF'),
-    generateDocumentOutline: checkType(options.generateDocumentOutline ?? false, 'boolean', 'generateDocumentOutline'),
-    ...pageSize
-  };
-
-  if (!this._printToPDF) {
-    throw new Error('Printing feature is disabled');
-  }
-
-  const prev = printToPDFQueues.get(this) ?? Promise.resolve();
-  const next = prev.catch(() => {}).then(() => this._printToPDF(printSettings));
-  printToPDFQueues.set(this, next);
-  next
-    .finally(() => {
-      if (printToPDFQueues.get(this) === next) printToPDFQueues.delete(this);
-    })
-    .catch(() => {});
-  return next;
+  return printToPDF(this, options);
 };
 
 // TODO(codebytere): deduplicate argument sanitization by moving rest of
