+++ shell/browser/api/electron_api_web_frame_main.h
@@ -17,6 +17,7 @@
 #include "gin/wrappable.h"
 #include "mojo/public/cpp/bindings/pending_receiver.h"
 #include "mojo/public/cpp/bindings/remote.h"
+#include "printing/buildflags/buildflags.h"
 #include "shell/browser/event_emitter_mixin.h"
 #include "shell/common/api/api.mojom.h"
 #include "shell/common/gc_plugin.h"
@@ -110,6 +111,9 @@ class WebFrameMain final : public gin::Wrappable<WebFrameMain>,
 
   v8::Local<v8::Promise> ExecuteJavaScript(gin::Arguments* args,
                                            const std::u16string& code);
+#if BUILDFLAG(ENABLE_PRINTING)
+  v8::Local<v8::Promise> PrintToPDF(const base::Value& settings);
+#endif
   void CopyVideoFrameAt(int x, int y);
   void SaveVideoFrameAs(int x, int y);
   bool Reload();
