+++ spec/api-web-contents-spec.ts
@@ -12,7 +12,6 @@ import {
 
 import { assert, expect } from 'chai';
 
-import * as cp from 'node:child_process';
 import { once } from 'node:events';
 import * as fs from 'node:fs';
 import * as http from 'node:http';
@@ -23,6 +22,7 @@ import { setTimeout } from 'node:timers/promises';
 import * as url from 'node:url';
 
 import { captureWithTabSourceId } from './lib/media-helpers';
+import { containsText, readPDF } from './lib/pdf-helpers';
 import { ifdescribe, defer, waitUntil, listen, ifit } from './lib/spec-helpers';
 import { cleanupWebContents, closeAllWindows } from './lib/window-helpers';
 
@@ -3832,45 +3832,8 @@ describe('webContents module', () => {
 
   ifdescribe(features.isPrintingEnabled())('printToPDF()', () => {
     let server: http.Server | null;
-    const readPDF = async (data: any) => {
-      const tmpDir = await fs.promises.mkdtemp(path.resolve(os.tmpdir(), 'e-spec-printtopdf-'));
-      const pdfPath = path.resolve(tmpDir, 'test.pdf');
-      await fs.promises.writeFile(pdfPath, data);
-      const pdfReaderPath = path.resolve(fixturesPath, 'api', 'pdf-reader.mjs');
-
-      const result = cp.spawn(process.execPath, [pdfReaderPath, pdfPath], {
-        stdio: 'pipe'
-      });
-
-      const stdout: Buffer[] = [];
-      const stderr: Buffer[] = [];
-      result.stdout.on('data', (chunk) => stdout.push(chunk));
-      result.stderr.on('data', (chunk) => stderr.push(chunk));
-
-      const [code, signal] = await new Promise<[number | null, NodeJS.Signals | null]>((resolve) => {
-        result.on('close', (code, signal) => {
-          resolve([code, signal]);
-        });
-      });
-      await fs.promises.rm(tmpDir, { force: true, recursive: true });
-      if (code !== 0) {
-        const errMsg = Buffer.concat(stderr).toString().trim();
-        console.error(`Error parsing PDF file, exit code was ${code}; signal was ${signal}, error: ${errMsg}`);
-      }
-      try {
-        return JSON.parse(Buffer.concat(stdout).toString().trim());
-      } catch (err) {
-        console.error('Error parsing PDF file:', err);
-        console.error('Raw output:', Buffer.concat(stdout).toString().trim());
-        throw err;
-      }
-    };
 
     let w: BrowserWindow;
-
-    const containsText = (items: any[], text: RegExp) => {
-      return items.some(({ str }: { str: string }) => str.match(text));
-    };
 
     beforeEach(() => {
       w = new BrowserWindow({
