+++ typings/internal-electron.d.ts
@@ -148,6 +148,7 @@ declare namespace Electron {
     _send(internal: boolean, channel: string, args: any): void;
     _sendInternal(channel: string, ...args: any[]): void;
     _postMessage(channel: string, message: any, transfer?: any[]): void;
+    _printToPDF(options: any): Promise<Buffer>;
     _lifecycleStateForTesting: string;
   }
